small optimization for OpenOCD
This makes most JTAG commands in write_memory_progbuf (riscv-013.c)
use output-only transfers. This means that the driver for the USB
blaster won't have to flush and wait for the reply that will be
discarded anyway. That way, ublast_buf_write can write the data in
batches of over 1000 bytes rather than 28 bytes at a time.

I had hoped that this would increase flashing speed but it doesn't.