don't override routeromen input for sub-flakes
This would use routeromen with pre-committed inputs (especially
nixpkgs) so any input overrides in the sub-flake would be ignored.
In particular, this would use the nixpkgs version from the routeromen
lock file for any use of nixpkgs inside routeromen. These should
follow the nixpkgs in the sub-flake (if the sub-flake says so).