disable speedtest to France
We can't do this because speedtest-cli doesn't support arbitrary IDs.
It has to be one from the list that it fetches so we are at the whim
of whomever generates the list for us.

We could get the URL from this list but speedtest-cli doesn't provide
any argument to pass this in:
https://www.speedtest.net/speedtest-servers-static.php