use monkeysphere to convert OpenPGP->OpenSSH
I was generating an additional SSH key as a fallback because gpg-agent
doesn't really seem to be too reliable. Now, we are using monkeysphere
to convert the auth subkey to OpenSSH format, which gives us the best
of both, i.e. only one SSH key and it will work with and without gpg-agent.