gk3v-pb: pin MongoDB package
The build is still ongoing and over 9 hours have passed. Yes, really.
I don't want to pin it but there is no way we can do this on every update.
Others are also running into this problem and the issue has been open
for some years. The only other option seem to be third-party caches
but I don't want to do that either.
-> MongoDB is network software but we don't want that instance to be
   available to anything but Omada anyway so maybe pinning is ok-ish......

https://discourse.nixos.org/t/pin-package-mongodb-to-stop-building-from-source/17452
https://github.com/NixOS/nixpkgs/issues/83884