omada-controller: 5.15.20.20 -> 6.0.0.24
This will need an upgrade of the database. They easiest way is to
remove it and import an autobackup.

This is (slightly) different from similar issues with Unifi because
the database is already in WiredTiger format but it is too old anyway.

The positive aspect is that the controller will talk to current
versions of MongoDB so we don't need the legacy version anymore.